move parse region to model region, and subdivide tests
